WIP initial attempt to run spec tests in browser.
