+++ package.json
@@ -21,6 +21,7 @@
   },
   "homepage": "https://github.com/sass/sass-spec#readme",
   "dependencies": {
+    "@esbuild-plugins/node-modules-polyfill": "^0.2.2",
     "@jest/types": "^27.2.5",
     "@types/data-urls": "^2.0.1",
     "@types/diff": "^5.0.1",
@@ -28,6 +29,7 @@
     "@types/intercept-stdout": "^0.1.0",
     "@types/jest": "^27.0.2",
     "@types/js-yaml": "^4.0.3",
+    "@types/karma": "^6.3.3",
     "@types/lodash": "^4.14.168",
     "@types/node": "^14.14.7",
     "@types/node-hrx": "^0.1.0",
@@ -36,15 +38,21 @@
     "data-urls": "^3.0.1",
     "del": "^6.0.0",
     "diff": "^5.0.0",
+    "esbuild": "^0.16.17",
     "file-lines-count": "^1.0.1",
     "fs-extra": "^10.0.0",
     "gts": "^3.1.0",
     "immutable": "^4.0.0",
     "intercept-stdout": "^0.1.2",
+    "jasmine-core": "^4.6.0",
     "jest": "^27.3.1",
     "jest-environment-node-single-context": "^27.2.0",
     "jest-extended": "^1.1.0",
     "js-yaml": "^4.1.0",
+    "karma": "^6.4.1",
+    "karma-chrome-launcher": "^3.1.1",
+    "karma-esbuild": "^2.2.5",
+    "karma-jasmine": "^5.1.0",
     "lodash": "^4.17.21",
     "node-hrx": "^0.1.0",
     "prettier": "^2.1.2",
